added error handling, created Space Model and implemented dependencies between entities, fixed booking update controller

+++ src/main/java/ch/zli/m223/m223project/Controller/UserController.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -15,7 +16,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import ch.zli.m223.m223project.Model.ApplicationUser;
 import ch.zli.m223.m223project.Repository.UserRepository;
@@ -38,35 +41,58 @@ public class UserController {
     PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
 
     @PostMapping(value = "/register", consumes = "application/json", produces = "application/json", name = "createUser")
+    @ResponseStatus(code = org.springframework.http.HttpStatus.CREATED, reason = "User created successfully!")
     public ApplicationUser createUser(@RequestBody ApplicationUser user) {
-        userRepo.save(
+        try{
+            userRepo.save(
                 new ApplicationUser(user.getPrename(), user.getSurname(), passwordEncoder.encode(user.getPassword()),
                         user.getUsername(), user.getRoles()));
 
-        return user;
+            return user;
+        } catch (Exception e){
+            throw new ResponseStatusException(HttpStatus.EXPECTATION_FAILED, "EXPECTATION FAILED(CODE 417)\n");
+        }
     }
 
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     @GetMapping(value = "/admin/users", name = "getAllUsers")
+    @ResponseStatus(code = HttpStatus.OK, reason = "Users found!")
     public List<ApplicationUser> manageUsers() {
-        return userRepo.findAll();
+        try {
+            return userRepo.findAll();
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No Users found!");
+        }
+        
     }
 
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     @GetMapping(value = "/admin/users/{id}", name = "getUserById")
+    @ResponseStatus(code = HttpStatus.FOUND, reason = "User found!")
     public Optional<ApplicationUser> findUserById(@PathVariable("id") Long id) {
-        return userRepo.findById(id);
+        Optional<ApplicationUser> user = userRepo.findById(id);
+        if (user.isPresent()) {
+            return userRepo.findById(id);
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No user found with id: " + id);        
     }
 
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     @DeleteMapping(value = "/admin/users/{id}", name = "deleteUserById")
+    @ResponseStatus(code = HttpStatus.OK, reason = "User deleted successfully!")
     public String deleteUserbyId(@PathVariable("id") Long id) {
-        userRepo.deleteById(id);
-        return "Deleted user with id: " + id;
+        Optional<ApplicationUser> user = userRepo.findById(id);
+        if (user.isPresent()) {
+            userRepo.deleteById(id);
+            return "Deleted user with id: " + id;
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No user found with id: " + id + " to delete!");
+        }
     }
 
     @PreAuthorize("hasRole('ROLE_ADMIN')")
     @PutMapping(value = "/admin/users/{id}", consumes = "application/json", produces = "application/json", name = "updateUserById")
+    @ResponseStatus(code = HttpStatus.OK, reason = "User updated successfully!")
     public String updateUserById(@PathVariable("id") Long id, @RequestBody ApplicationUser user) {
         Optional<ApplicationUser> userOptional = userRepo.findById(id);
 
@@ -79,16 +105,21 @@ public class UserController {
             userToUpdate.setRoles(user.getRoles());
             userRepo.save(userToUpdate);
             return "Updated user with id: " + id;
-        } else {
-            return "User not found";
-        }
+        } 
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No user found with id: " + id + " to update!");
     }
 
     @PostMapping(value = "/token", name = "getTokenForUser")
+    @ResponseStatus(code = HttpStatus.OK, reason = "Token generated successfully!")
     public String token(Authentication authentication) {
-        LOG.debug("Token requested for user {}", authentication.getName());
-        String token = tokenService.generateToken(authentication);
-        LOG.debug("Token generated: {}", token);
-        return token;
+        try{
+                    LOG.debug("Token requested for user {}", authentication.getName());
+            String token = tokenService.generateToken(authentication);
+            LOG.debug("Token generated: {}", token);
+            return token;
+        } catch (Exception e) {
+            throw new ResponseStatusException(HttpStatus.EXPECTATION_FAILED, "EXPECTATION FAILED(CODE 417)\n");
+        }
+
     }
 }
